refactor: create common file for db methods and hide helpers
