+++ src/db/methods/__init__.py
@@ -2,13 +2,13 @@ from . import (
     users,
     domains,
     organizations,
-    helpers,
+    common,
 )
 
 
 __all__ = [
     "users",
     "domains",
-    "helpers",
+    "common",
     "organizations",
 ]
