+++ src/db/methods/helpers.py
@@ -7,7 +7,6 @@ from db.types import types
 from .collections import internal_counters, users, organizations, domains
 
 
-
 def insert_with_auto_increment_id(
     collection: Collection[Any],
     document: dict[str, Any],
@@ -30,19 +29,7 @@ def insert_with_auto_increment_id(
             ):
                 raise e
 
-def get_object_by_id[T: types.DocumentWithIntId](
-    object_id: int, object_type: type[T], session: ClientSession | None = None,
-) -> T | None:
-    if (obj := _get_related_collection(object_type).find_one({"_id": object_id}, session=session)) is None:
-        return None
-    return object_type(**obj)
-
-def check_object_existence(
-    object_id: int, object_type: type[types.DocumentWithIntId], session: ClientSession | None = None,
-) -> bool:
-    return _get_related_collection(object_type).count_documents({"_id": object_id}, session=session) > 0
-
-def _get_related_collection(object_type: type[types.HasCollection]) -> Collection:
+def get_related_collection(object_type: type[types.HasCollection]) -> Collection:
     match object_type:
         case types.User:
             return users
