+++ src/routers/organizations.py
@@ -11,7 +11,7 @@ router = APIRouter()
 
 @router.get("/get", response_model=SuccessfulResponse[RS.organizations.get])
 async def get(request: RQ.organizations.get = Depends(), _current_user: types.User = Depends(get_current_user)):
-    if (organization := methods.helpers.get_object_by_id(request.id, types.Organization)) is None:
+    if (organization := methods.common.get_object_by_id(request.id, types.Organization)) is None:
         raise ErrorResponse(code=ErrorCodes.ENTITY_NOT_FOUND)
 
     return organization
